Crate rusoto_firehose

Amazon Kinesis Data Firehose API Reference
Amazon Kinesis Data Firehose is a fully managed service that delivers real-time streaming data to destinations such as Amazon Simple Storage Service (Amazon S3), Amazon Elasticsearch Service (Amazon ES), Amazon Redshift, and Splunk.

If you're using the service, you're probably looking for KinesisFirehoseClient and KinesisFirehose.

Structs

BufferingHints
Describes hints for the buffering to perform before delivering data to the destination. These options are treated as hints, and therefore Kinesis Data Firehose might choose to use different values when it is optimal.

CloudWatchLoggingOptions
Describes the Amazon CloudWatch logging options for your delivery stream.

CopyCommand
Describes a COPY command for Amazon Redshift.

CreateDeliveryStreamInput

CreateDeliveryStreamOutput

DataFormatConversionConfiguration
Specifies that you want Kinesis Data Firehose to convert data from the JSON format to the Parquet or ORC format before writing it to Amazon S3. Kinesis Data Firehose uses the serializer and deserializer that you specify, in addition to the column information from the AWS Glue table, to deserialize your input data from JSON and then serialize it to the Parquet or ORC format. For more information, see Kinesis Data Firehose Record Format Conversion.

DeleteDeliveryStreamInput

DeleteDeliveryStreamOutput

DeliveryStreamDescription
Contains information about a delivery stream.

DeliveryStreamEncryptionConfiguration
Indicates the server-side encryption (SSE) status for the delivery stream.

DescribeDeliveryStreamInput

DescribeDeliveryStreamOutput

Deserializer
The deserializer you want Kinesis Data Firehose to use for converting the input data from JSON. Kinesis Data Firehose then serializes the data to its final format using the Serializer. Kinesis Data Firehose supports two types of deserializers: the Apache Hive JSON SerDe and the OpenX JSON SerDe.

DestinationDescription
Describes the destination for a delivery stream.

ElasticsearchBufferingHints
Describes the buffering to perform before delivering data to the Amazon ES destination.

ElasticsearchDestinationConfiguration
Describes the configuration of a destination in Amazon ES.

ElasticsearchDestinationDescription
The destination description in Amazon ES.

ElasticsearchDestinationUpdate
Describes an update for a destination in Amazon ES.

ElasticsearchRetryOptions
Configures retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon ES.

EncryptionConfiguration
Describes the encryption for a destination in Amazon S3.

ExtendedS3DestinationConfiguration
Describes the configuration of a destination in Amazon S3.

ExtendedS3DestinationDescription
Describes a destination in Amazon S3.

ExtendedS3DestinationUpdate
Describes an update for a destination in Amazon S3.

HiveJsonSerDe
The native Hive / HCatalog JsonSerDe. Used by Kinesis Data Firehose for deserializing data, which means converting it from the JSON format in preparation for serializing it to the Parquet or ORC format. This is one of two deserializers you can choose, depending on which one offers the functionality you need. The other option is the OpenX SerDe.

InputFormatConfiguration
Specifies the deserializer you want to use to convert the format of the input data.

KMSEncryptionConfig
Describes an encryption key for a destination in Amazon S3.

KinesisFirehoseClient
A client for the Firehose API.

KinesisStreamSourceConfiguration
The stream and role Amazon Resource Names (ARNs) for a Kinesis data stream used as the source for a delivery stream.

KinesisStreamSourceDescription
Details about a Kinesis data stream used as the source for a Kinesis Data Firehose delivery stream.

ListDeliveryStreamsInput

ListDeliveryStreamsOutput

ListTagsForDeliveryStreamInput

ListTagsForDeliveryStreamOutput

OpenXJsonSerDe
The OpenX SerDe. Used by Kinesis Data Firehose for deserializing data, which means converting it from the JSON format in preparation for serializing it to the Parquet or ORC format. This is one of two deserializers you can choose, depending on which one offers the functionality you need. The other option is the native Hive / HCatalog JsonSerDe.

OrcSerDe
A serializer to use for converting data to the ORC format before storing it in Amazon S3. For more information, see Apache ORC.

OutputFormatConfiguration
Specifies the serializer that you want Kinesis Data Firehose to use to convert the format of your data before it writes it to Amazon S3.

ParquetSerDe
A serializer to use for converting data to the Parquet format before storing it in Amazon S3. For more information, see Apache Parquet.

ProcessingConfiguration
Describes a data processing configuration.

Processor
Describes a data processor.

ProcessorParameter
Describes the processor parameter.

PutRecordBatchInput

PutRecordBatchOutput

PutRecordBatchResponseEntry
Contains the result for an individual record from a PutRecordBatch request. If the record is successfully added to your delivery stream, it receives a record ID. If the record fails to be added to your delivery stream, the result includes an error code and an error message.

PutRecordInput

PutRecordOutput

Record
The unit of data in a delivery stream.

RedshiftDestinationConfiguration
Describes the configuration of a destination in Amazon Redshift.

RedshiftDestinationDescription
Describes a destination in Amazon Redshift.

RedshiftDestinationUpdate
Describes an update for a destination in Amazon Redshift.

RedshiftRetryOptions
Configures retry behavior in case Kinesis Data Firehose is unable to deliver documents to Amazon Redshift.

S3DestinationConfiguration
Describes the configuration of a destination in Amazon S3.

S3DestinationDescription
Describes a destination in Amazon S3.

S3DestinationUpdate
Describes an update for a destination in Amazon S3.

SchemaConfiguration
Specifies the schema to which you want Kinesis Data Firehose to configure your data before it writes it to Amazon S3.

Serializer
The serializer that you want Kinesis Data Firehose to use to convert data to the target format before writing it to Amazon S3. Kinesis Data Firehose supports two types of serializers: the ORC SerDe and the Parquet SerDe.

SourceDescription
Details about a Kinesis data stream used as the source for a Kinesis Data Firehose delivery stream.

SplunkDestinationConfiguration
Describes the configuration of a destination in Splunk.

SplunkDestinationDescription
Describes a destination in Splunk.

SplunkDestinationUpdate
Describes an update for a destination in Splunk.

SplunkRetryOptions
Configures retry behavior in case Kinesis Data Firehose is unable to deliver documents to Splunk, or if it doesn't receive an acknowledgment from Splunk.

StartDeliveryStreamEncryptionInput

StartDeliveryStreamEncryptionOutput

StopDeliveryStreamEncryptionInput

StopDeliveryStreamEncryptionOutput

Tag
Metadata that you can assign to a delivery stream, consisting of a key-value pair.

TagDeliveryStreamInput

TagDeliveryStreamOutput

UntagDeliveryStreamInput

UntagDeliveryStreamOutput

UpdateDestinationInput

UpdateDestinationOutput

Enums

CreateDeliveryStreamError
Errors returned by CreateDeliveryStream

DeleteDeliveryStreamError
Errors returned by DeleteDeliveryStream

DescribeDeliveryStreamError
Errors returned by DescribeDeliveryStream

ListDeliveryStreamsError
Errors returned by ListDeliveryStreams

ListTagsForDeliveryStreamError
Errors returned by ListTagsForDeliveryStream

PutRecordBatchError
Errors returned by PutRecordBatch

PutRecordError
Errors returned by PutRecord

StartDeliveryStreamEncryptionError
Errors returned by StartDeliveryStreamEncryption

StopDeliveryStreamEncryptionError
Errors returned by StopDeliveryStreamEncryption

TagDeliveryStreamError
Errors returned by TagDeliveryStream

UntagDeliveryStreamError
Errors returned by UntagDeliveryStream

UpdateDestinationError
Errors returned by UpdateDestination

Traits

KinesisFirehose
Trait representing the capabilities of the Firehose API. Firehose clients implement this trait.