Struct rusoto_firehose::CreateDeliveryStreamInput

pub struct CreateDeliveryStreamInput {
    pub delivery_stream_name: String,
    pub elasticsearch_destination_configuration: Option<ElasticsearchDestinationConfiguration>,
    pub extended_s3_destination_configuration: Option<ExtendedS3DestinationConfiguration>,
    pub redshift_destination_configuration: Option<RedshiftDestinationConfiguration>,
}

Fields
delivery_stream_name: String

The name of the delivery stream. This name must be unique per AWS account in the same region. You can have multiple delivery streams with the same name if they are in different accounts or different regions.

elasticsearch_destination_configuration: Option<ElasticsearchDestinationConfiguration>

The destination in Amazon ES. You can specify only one destination.

extended_s3_destination_configuration: Option<ExtendedS3DestinationConfiguration>

The destination in Amazon S3. You can specify only one destination.

redshift_destination_configuration: Option<RedshiftDestinationConfiguration>

The destination in Amazon Redshift. You can specify only one destination.