Struct rusoto_firehose::ElasticsearchRetryOptions

pub struct ElasticsearchRetryOptions {
    pub duration_in_seconds: Option<i64>,
}

Configures retry behavior in the event that Firehose is unable to deliver documents to Amazon ES.

Fields
duration_in_seconds: Option<i64>

After an initial failure to deliver to Amazon ES, the total amount of time during which Firehose re-attempts delivery (including the first attempt). After this time has elapsed, the failed documents are written to Amazon S3. Default value is 300 seconds (5 minutes). A value of 0 (zero) results in no retries.