Struct rusoto_firehose::KinesisFirehoseClient

pub struct KinesisFirehoseClient { /* fields omitted */ }

A client for the Firehose API.

Methods

impl KinesisFirehoseClient

pub fn new(region: Region) -> KinesisFirehoseClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> KinesisFirehoseClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,