Struct rusoto_firehose::RedshiftDestinationUpdate

pub struct RedshiftDestinationUpdate {
    pub cloud_watch_logging_options: Option<CloudWatchLoggingOptions>,
    pub cluster_jdbcurl: Option<String>,
    pub copy_command: Option<CopyCommand>,
    pub password: Option<String>,
    pub processing_configuration: Option<ProcessingConfiguration>,
    pub retry_options: Option<RedshiftRetryOptions>,
    pub role_arn: Option<String>,
    pub s3_backup_mode: Option<String>,
    pub s3_backup_update: Option<S3DestinationUpdate>,
    pub s3_update: Option<S3DestinationUpdate>,
    pub username: Option<String>,
}

Describes an update for a destination in Amazon Redshift.

Fields
cloud_watch_logging_options: Option<CloudWatchLoggingOptions>

The CloudWatch logging options for your delivery stream.

cluster_jdbcurl: Option<String>

The database connection string.

copy_command: Option<CopyCommand>

The COPY command.

password: Option<String>

The user password.

processing_configuration: Option<ProcessingConfiguration>

The data processing configuration.

retry_options: Option<RedshiftRetryOptions>

The retry behavior in the event that Firehose is unable to deliver documents to Amazon Redshift. Default value is 3600 (60 minutes).

role_arn: Option<String>

The ARN of the AWS credentials.

s3_backup_mode: Option<String>

The Amazon S3 backup mode.

s3_backup_update: Option<S3DestinationUpdate>

The Amazon S3 destination for backup.

s3_update: Option<S3DestinationUpdate>

The Amazon S3 destination.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationUpdate.S3Update because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

username: Option<String>

The name of the user.