Struct rusoto_firehose::SplunkDestinationDescription

pub struct SplunkDestinationDescription {
    pub cloud_watch_logging_options: Option<CloudWatchLoggingOptions>,
    pub hec_acknowledgment_timeout_in_seconds: Option<i64>,
    pub hec_endpoint: Option<String>,
    pub hec_endpoint_type: Option<String>,
    pub hec_token: Option<String>,
    pub processing_configuration: Option<ProcessingConfiguration>,
    pub retry_options: Option<SplunkRetryOptions>,
    pub s3_backup_mode: Option<String>,
    pub s3_destination_description: Option<S3DestinationDescription>,
}

Describes a destination in Splunk.

Fields
cloud_watch_logging_options: Option<CloudWatchLoggingOptions>

The CloudWatch logging options for your delivery stream.

hec_acknowledgment_timeout_in_seconds: Option<i64>

The amount of time that Kinesis Firehose waits to receive an acknowledgment from Splunk after it sends it data. At the end of the timeout period Kinesis Firehose either tries to send the data again or considers it an error, based on your retry settings.

hec_endpoint: Option<String>

The HTTP Event Collector (HEC) endpoint to which Kinesis Firehose sends your data.

hec_endpoint_type: Option<String>

This type can be either "Raw" or "Event".

hec_token: Option<String>

This is a GUID you obtain from your Splunk cluster when you create a new HEC endpoint.

processing_configuration: Option<ProcessingConfiguration>

The data processing configuration.

retry_options: Option<SplunkRetryOptions>

The retry behavior in case Kinesis Firehose is unable to deliver data to Splunk or if it doesn't receive an acknowledgment of receipt from Splunk.

s3_backup_mode: Option<String>

Defines how documents should be delivered to Amazon S3. When set to FailedDocumentsOnly, Kinesis Firehose writes any data that could not be indexed to the configured Amazon S3 destination. When set to AllDocuments, Kinesis Firehose delivers all incoming records to Amazon S3, and also writes failed documents to Amazon S3. Default value is FailedDocumentsOnly.

s3_destination_description: Option<S3DestinationDescription>

The Amazon S3 destination.>