Crate rusoto_firehose

Amazon Kinesis Firehose API Reference
Amazon Kinesis Firehose is a fully-managed service that delivers real-time streaming data to destinations such as Amazon Simple Storage Service (Amazon S3), Amazon Elasticsearch Service (Amazon ES), and Amazon Redshift.

If you're using the service, you're probably looking for KinesisFirehoseClient and KinesisFirehose.

Structs

BufferingHints
Describes hints for the buffering to perform before delivering data to the destination. Please note that these options are treated as hints, and therefore Firehose may choose to use different values when it is optimal.

CloudWatchLoggingOptions
Describes the CloudWatch logging options for your delivery stream.

CopyCommand
Describes a COPY command for Amazon Redshift.

CreateDeliveryStreamInput

CreateDeliveryStreamOutput

DeleteDeliveryStreamInput

DeleteDeliveryStreamOutput

DeliveryStreamDescription
Contains information about a delivery stream.

DescribeDeliveryStreamInput

DescribeDeliveryStreamOutput

DestinationDescription
Describes the destination for a delivery stream.

ElasticsearchBufferingHints
Describes the buffering to perform before delivering data to the Amazon ES destination.

ElasticsearchDestinationConfiguration
Describes the configuration of a destination in Amazon ES.

ElasticsearchDestinationDescription
The destination description in Amazon ES.

ElasticsearchDestinationUpdate
Describes an update for a destination in Amazon ES.

ElasticsearchRetryOptions
Configures retry behavior in the event that Firehose is unable to deliver documents to Amazon ES.

EncryptionConfiguration
Describes the encryption for a destination in Amazon S3.

ExtendedS3DestinationConfiguration
Describes the configuration of a destination in Amazon S3.

ExtendedS3DestinationDescription
Describes a destination in Amazon S3.

ExtendedS3DestinationUpdate
Describes an update for a destination in Amazon S3.

KMSEncryptionConfig
Describes an encryption key for a destination in Amazon S3.

KinesisFirehoseClient
A client for the Firehose API.

ListDeliveryStreamsInput

ListDeliveryStreamsOutput

ProcessingConfiguration
Describes a data processing configuration.

Processor
Describes a data processor.

ProcessorParameter
Describes the processor parameter.

PutRecordBatchInput

PutRecordBatchOutput

PutRecordBatchResponseEntry
Contains the result for an individual record from a PutRecordBatch request. If the record is successfully added to your delivery stream, it receives a record ID. If the record fails to be added to your delivery stream, the result includes an error code and an error message.

PutRecordInput

PutRecordOutput

Record
The unit of data in a delivery stream.

RedshiftDestinationConfiguration
Describes the configuration of a destination in Amazon Redshift.

RedshiftDestinationDescription
Describes a destination in Amazon Redshift.

RedshiftDestinationUpdate
Describes an update for a destination in Amazon Redshift.

RedshiftRetryOptions
Configures retry behavior in the event that Firehose is unable to deliver documents to Amazon Redshift.

S3DestinationConfiguration
Describes the configuration of a destination in Amazon S3.

S3DestinationDescription
Describes a destination in Amazon S3.

S3DestinationUpdate
Describes an update for a destination in Amazon S3.

UpdateDestinationInput

UpdateDestinationOutput

Enums

CreateDeliveryStreamError
Errors returned by CreateDeliveryStream

DeleteDeliveryStreamError
Errors returned by DeleteDeliveryStream

DescribeDeliveryStreamError
Errors returned by DescribeDeliveryStream

ListDeliveryStreamsError
Errors returned by ListDeliveryStreams

PutRecordBatchError
Errors returned by PutRecordBatch

PutRecordError
Errors returned by PutRecord

UpdateDestinationError
Errors returned by UpdateDestination

Traits

KinesisFirehose
Trait representing the capabilities of the Firehose API. Firehose clients implement this trait.